==> 3.1 Estadisticas totales.ipynb <==
import pandas as pd
import os
from funciones_auxiliares import *
import warnings
warnings.filterwarnings('ignore')

# 3.1 *Estadisticas totales*

Las *estadisticas totales*, también conocidas como estadísticas acumuladas, es la información que representa las estadísticas obtenidas durante un jugador en un periodo de tiempo como puede ser un número de partidos específicos, una temporada o varias. En este notebook se buscará dar explicación de manera entendible y visualizada a todo el procesamiento realizado para la construcción de estas estadísticas a través de los siguientes puntos:
* **1. Base teórica de la construcción de las estadísticas.** Se realizará la explicación teórica de los pasos realizados para la construcción de este tipo de estadísticas.
* **2. Procesamiento de los datos.** Se realizará la explicación guiada del procesamiento de los datos para la construcción de las estadísticas totales.

A través de esos puntos se realizará la comprensión y entendimiento del trabajo realizado para la construcción de las *estadísticas totales*.

## 1. Base teórica de la construcción de las estadísticas

Como se ha comentado al inicio de este notebook, las *estadísticas totales*, es el tipo de información que representa las estadísticas obtenidas por un jugador a lo largo de un periodo de tiempo. Como ha sido explicado en el notebook [Descarga de *Box Score*.](2.3%20Box%20Score), han sido descargados todos los *Box Scores* de los partidos disputados entre las temporadas *2019-2020*, *2020-2021*, *2021-2022* y *2022-2023* siendo este el motivo que para la construcción de las *estadísticas totales* sean utilizado estos datos.

Para la construcción de este tipo de estadísticas, se realizará el sumatorio de todos los apartados estadísticos descargados para cada jugador en una temporada y un equipo específico teniendo así todas las estadísticas obtenidas por un jugador durante cada una de las temporadas dependiendo del equipo en el que hayan participado en la competición. Para la creación de este tipo de estadística ha sido desarrollado el código [03_Procesamiento_PBP](03_Procesamiento_PBP.py) que permite el procesamiento de cada uno de los *Box Score* descargados.

# 2. Procesamiento de los datos

Para poder comprender el procesamiento realizado sobre los datos, se pasa a realizar la carga de todas las estadísticas obtenidas por el jugador *Sergio Llull* durante la temporada *2022-2023* en el equipo *Real Madrid*:

# Ruta para acceder a la información del Real Madrid en la temporada 2022
ruta_equipo = 'E:/TFM/02. Datos/ACB/EQUIPOS/2022/Real Madrid'

# Ruta para acceder al fichero de estadísticas totales del Real Madrid y sus rivales
fichero_total_2022 = 'E:/TFM/02. Datos/ACB/ESTADISTICAS/2022/Real Madrid/01_ESTADISTICAS_TOTALES_EQUIPO.csv'
fichero_total_rival_2022 = 'E:/TFM/02. Datos/ACB/ESTADISTICAS/2022/Real Madrid/01_ESTADISTICAS_TOTALES_RIVAL.csv'

# Ruta de estadísticas
ruta_estadisticas = 'E:/TFM/02. Datos/ACB/ESTADISTICAS'

df_jugador = pd.DataFrame()
l_partidos = os.listdir(ruta_equipo)
equipo = 'Real Madrid'
jugador = 'S. Llull'

for partido in l_partidos:
    ruta_dir_partido = ruta_equipo + '/' + partido
    l_elementos = os.listdir(ruta_dir_partido)
    
    # Se realiza la selección de los ficheros y la carga de ellos
    nombre_fichero_equipo = [elem for elem in l_elementos if equipo in elem and 'VS' not in elem and 'CARTA_TIRO' not in elem][0]
    nombre_fichero_rival = [elem for elem in l_elementos if equipo not in elem and 'VS' not in elem and 'CARTA_TIRO' not in elem][0]
    
    df_fichero_equipo = pd.read_csv(ruta_dir_partido + '/' + nombre_fichero_equipo, sep=';')
    df_fichero_rival = pd.read_csv(ruta_dir_partido + '/' + nombre_fichero_rival, sep=';')
    
    # Se realiza la selección del registro del jugador y del equipo rival
    df_fichero_jugador = df_fichero_equipo[df_fichero_equipo['JUGADOR'] == jugador]
    df_jugador = pd.concat([df_jugador,df_fichero_jugador])

print("################################################################################")
print("# ESTADISTICAS EN CADA PARTIDO JUGADO DE SERGIO LULL EN LA TEMPORADA 2022-2023")
print("################################################################################") 
print("Nº de columnas:", len(df_jugador.columns))
print("Nº de registros:", len(df_jugador))
print("\n",df_jugador.to_string(index=False, col_space=4))    

Como puede observarse, y como se vío en la explicación de la información descargada, existe el campo *Unnamed:0* el cual se encuentra vacío además de contar con campos como *MIN*, *T2*, *T2%*, *T3*, *T3%*, *TL* y *TL%* que se encuentran en formato que no permite el sumatorio de estos directamente. Es por este motivo que ha sido realizado un procesamineto de los datos el cual cosiste en la eliminación de la columna *Unnamed: 0*, la conversión del campo *MIN* únicamente a minutos, la eliminación de los campos que representan los porcentajes de tiro y la división en tiros lanzados y tiros anotados de cada uno de los campos de tiro. Teniendo esto presente, y tras realizar el procesamiento, los datos quedan de la siguiente manera:

print("################################################################################")
print("# PROCESAMIENTO DE ESTADISTICAS DE SERGIO LULL EN LA TEMPORADA 2022-2023")
print("################################################################################") 
df_jugador_precesado = procesar_BS(df_jugador)
print("Nº de columnas:", len(df_jugador_precesado.columns))
print("Nº de registros:", len(df_jugador_precesado))
print("\n",df_jugador_precesado.to_string(index=False, col_space=4)) 

Como puede observarse, han sido añadiddas las columnas *T2A*, *T2L*, *T3A*, *T3L*, *TLA* y *TLL* además de eliminar la columna con elementos vacíos y convertir la variable *MIN*. A través de esta información se pasa a construir las estadísticas totales del jugador a lo largo de la temporada mediante el sumatorio de cada uno de los registros:

print("################################################################################")
print("# SUMATORIO DE ESTADISTICAS DE SERGIO LULL EN LA TEMPORADA 2022-2023")
print("################################################################################")
df_jugador_agrupado = df_jugador_precesado.groupby(by=['NO.','JUGADOR']).sum().reset_index() 
print("Nº de columnas:", len(df_jugador_agrupado.columns))
print("Nº de registros:", len(df_jugador_agrupado))
print("\n",df_jugador_agrupado.to_string(index=False, col_space=4)) 

Como puede observarse, tras la aplicación del sumatorio, se obtienen la estadísticas del jugador a lo largo de la temporada. Además de las estadísticas totales de cada jugador, también han sido construidas las estadísticas totales de cada uno de los equipos y de sus rivales las cuales serán utilizadas para la construcción de futuras estadísticas:

# Se realiza la carga de los ficheros de estadísticas totales
df_totales_2022 = pd.read_csv(fichero_total_2022, sep = ';')
df_totales_rival_2022 = pd.read_csv(fichero_total_rival_2022, sep = ';')

print("##################################################################################")
print("# ESTADISTICAS TOTALES DE LOS JUGADORES DEL REAL MADRID EN LA TEMPORADA 2022-2023")
print("##################################################################################")
print("Nº de columnas:", len(df_totales_2022.columns))
print("Nº de registros:", len(df_totales_2022))
print("\n",df_totales_2022.to_string(index=False, col_space=4)) 

print("\n##################################################################################")
print("# ESTADISTICAS TOTALES DE LOS RIVALES DEL REAL MADRID EN LA TEMPORADA 2022-2023")
print("##################################################################################")
print("Nº de columnas:", len(df_totales_rival_2022.columns))
print("Nº de registros:", len(df_totales_rival_2022))
print("\n",df_totales_rival_2022.to_string(index=False, col_space=4))
