==> 3.4 Estadísticas avanzadas individuales.ipynb <==
import pandas as pd
import os
from funciones_auxiliares import *
import warnings
warnings.filterwarnings('ignore')

# 3.4 *Estadisticas avanzadas individuales*

Las *estadísticas avnzadas individuales* son el tipo de estadística definida por *Dean Oliver* en su libro *Basketball on Paper* las cuales permiten medir diferentes aspectos de los jugadores de baloncesto a partir de las *estadísticas totales* en un periodo de tiempo. En este notebook se buscará dar explicación de manera entendible y visualizada a todo el procesamiento realizado para la construcción de estas estadísticas a través de los siguientes puntos:
* **1. Base teórica de la construcción de las estadísticas.** Se realizará la explicación teórica de los pasos realizados para la construcción de este tipo de estadísticas.
* **2. Procesamiento de los datos.** Se realizará la explicación guiada del procesamiento de los datos para la construcción de las estadísticas totales.

A través de esos puntos se realizará la comprensión y entendimiento del trabajo realizado para la construcción de las *estadísticas totales*.

## 1. Base teórica de la construcción de las estadísticas

Como ha sido comentado al inicio de este notebook, las *estadísticas avanzadas individuales* fueron diseñadas por *Dean Oliver* en su libro *Baskeball on Paper* en el cual se definieron diferentes métricas las cuales permiten saber el desempeño de los jugadores de baloncesto en función de su estadísticas individuales totales. Estas estadísticas permite mostrar el desempeño de los jugadores de baloncesto dentro de una pista ya que representan cada una de las estadísticas teniendo en cuenta el contexto en el que las realizó el jugador. Es por este motivo que para la construcción de las *estadísticas avanzadas individuales* han sido utilizadas las estadísticas totales contruidas en el apartado [Construcción de las *estadísticas totales*.](3.1%20Estadisticas%20totales) tanto de jugadores como de equipos. Mediante este tipo de estadísticas será construidas las siguientes estadísticas avanzadas:
* *eFG%  - Effective Field Goal Percentage.* El *eFG%* conocida como *porcentaje de eficiencia del tiro*, es la métrica que permite conocer como de bunos han sido los tiros realizados por el jugador. Esta métrica es calcula a patir de la fórmula *eFG% = (Tiros anotados + 0.5 * Tiros de tres anotados) / Tiros lanzdaos*.
* *3Pr - Three-Point rating.* El *3Pr* conocida como *volumen de tiros de tres lanzados*, es la métrica que permite conocer el porcentaje de tiros de tres que ha sido lanzado por el jugador con respecto a su tiros totales. Est métrica es calculada a partir de la fórmula *3Pr = Tiros de tres lanzados / Tiros lanzados.
* *FTr - Free Throw rating.* El *Ftr* conocida como *volumen de tiros libres* es la métrica que permite conocer el volumen de tiros libres anotados con respecto a los tiros lanzados por el jugador. Esta métrica es calculada a partir de la fórmula *FTr = Tiros libres anotados / Tiros de campo lanzados.
* *ORB% -  Offensive Rebound Percentage.* El *ORB%* tambien conocida por *Volumen de rebotes ofensivos* es la métrica que permite conocer cuantos rebotes ofensivos del total de rebotes diponibles fue capaz de capturar el jugador. Esta métrica es calculada a partir de la fórmula *ORB% = (Rebotes ofensivos capturados por el jugador * (Minutos jugados por el equipo / 5)) / (Minutos jugados por el jugador * (Rebotes ofensivos del equipo + Rebotes defensivos del equipo rival))*.
* *DRB% -  Defensive Rebound Percentage.* El *DRB%* también conocida por *volumen de rebotes defensivos* es la métrica que permite conocer cuantos rebotes defensivos del total de rebotes dispones fue capaz de capturar el jugador. Esta métrica es calculada a a partir de la fórmula *DRB% = (Rebotes defensivos capturados por el jugador * (Minutos jugados por el equipo / 5)) / (Minutos jugados por el jugador * (Rebotes defensivos del equipo + Rebotes ofensivos del equipos rival))*.
* *STL% - Steal Percentage.* El *STL%* también conocida como *volumen de recuperaciones realizadas* es la métrica que permite conocer cuantos balones fueron robados por el jugador con respecto al total de balones robados por el equipo. Esta métrica es calculada a partir de la fórmula *STL% = (Robos realizados por el jugador * (Minutos jugados por el equipo / 5)) / (Minutos jugados por el jugador * Posesiones realizadas por el rival)*.
* *BLK% - Block Percentage.* El *BLK%* también conocida como *volumen de tapones realizados* es la métrica que permite conocer cuanto tiros del oponente fueron taponados con respecto al volumen total de tiros. Esta métrica es calculada a partir de la fórmula *BLK% = Tapones realizados por el jugador * (Minutos jugados por el equipo / 5)) / (Minutos jugados por el jugador * (Tiros de campo anotados por el rival - Tiros de tres anotados por el rival)*.
* *TOV% - Turnover Percentage.* El *TOV%* también conocida como *volumen de pérdidas realizadas* es la métrica que permite conocer cuantos balonces fueron perdidos por el jugador con respecto al total de balones perdidos del equipo. Esta métrica es calculada a partir de la fórmula *TOV% = Pérdidas realizadas por el jugador / (Tiros de campo anotados por el jugador + 0.44 * Tiros libres anotados por el jugador + Pérdidas realizadas por el jugador)*.
* *AST% - Assist Percentage.* El *AST%* también conocida como *volumen de asistencias realizadas* es la metrica que permite conoces el volumen de asistencias realizadas por el jugadr con respecto a los tiros realizados del equipo. Esta métrica es calculada a apartir de la fórmula *AST% = Asistencias del jugador / (((Minutos jugados por el jugador / (Minutos jugados por el equipo / 5)) * Tiros realizados por el equipo) - Tiros realizados por el jugador*.
* *USG% - Usage Percentage.* El *USG%* también conocida como *porcentaje de uso* es la métrica que permite conocer cuantas posesiones del equipo fueron finalizadas por el jugador ya sea por tiro, pérdida o falta recibida de tiro. Esta métrica es calculada a partir de la fórmula *USG% = (Tiros de campo lanzados por el jugador + 0.44 * Tiros libres lanzados por el jugador + Pérdidas realizadas por el jugador) * (Minutos jugados por el equipo / 5)) / (Minutos jugados por el jugador * (Tiros de campo lanzados por el jugador + 0.44 * Tiros libres lanzados por el equipo + Péridas realizadas por el equipo)*.
* *STOP%.*

A partir de estas estadísticas han sido creadas las estadísticas individuales de los jugadores. Para ello, ha sido creado el script de código [04_Procesamiento_ESTADISTICA_AVANZADA](04_Procesamiento_ESTADISTICA_AVANZADA.py) con el cual se realiza la construcción de las estadísticas avanzadas para cada jugador en cada una de las temporadas para las que han sido construidas las *estadísticas totales*.

# 2. Procesamiento de los datos

Como ha sido explicado en el apartado anterior, para el cálculo de las *estadísticas avanzadas individuales* es necesario tener la información por un lado de las estadísticas totales realizadas por los jugadores, las estadísticas totales de los equipos y las estadísticas totales de los rivales a lo que se haya enfrentado el equipo. Es por este motivo que para la construcción de este tipo de estadística se realiza la carga de todas las estadísticas totales del Real Madrid durante la temporada *2022-2023*:

# Ruta para acceder al fichero de estadísticas totales del Real Madrid
fichero_total_jugadores_2022 = 'E:/TFM/02. Datos/ACB/ESTADISTICAS/2022/Real Madrid/01_ESTADISTICAS_TOTALES_EQUIPO.csv'

# Ruta para acceder al fichero de estadísticas totales de los rivales del Real Madrid
fichero_rival_2022 = 'E:/TFM/02. Datos/ACB/ESTADISTICAS/2022/Real Madrid/01_ESTADISTICAS_TOTALES_RIVAL.csv'

df_totales_2022 = pd.read_csv(fichero_total_jugadores_2022, sep = ';')
df_totales_rival_2022 = pd.read_csv(fichero_rival_2022, sep = ';')

print("##################################################################################")
print("# ESTADISTICAS TOTALES DE LOS JUGADORES DEL REAL MADRID EN LA TEMPORADA 2022-2023")
print("##################################################################################")
print(df_totales_2022.to_string(index=False, col_space=4)) 

print("\n##################################################################################")
print("# ESTADISTICAS TOTALES DEL REAL MADRID EN LA TEMPORADA 2022-2023")
print("##################################################################################")
df_totales_2022['EQUIPO'] = 'EQUIPO'
df_equipo_total = df_totales_2022.drop(['NO.','JUGADOR'],axis=1).groupby('EQUIPO').sum().reset_index()
print(df_equipo_total.to_string(index=False, col_space=4)) 

print("\n##################################################################################")
print("# ESTADISTICAS TOTALES DE LOS RIVALES DEL REAL MADRID EN LA TEMPORADA 2022-2023")
print("##################################################################################")
print(df_totales_rival_2022.to_string(index=False, col_space=4)) 

A partir de las estadísticas totales mostradas y la aplicación de las fórmulas definidas en el primer apartado se pasa a mostrar la *estadísticas totales avanzadas* de los jugadores del Real Madrid durante la temporada *2022-2023*:

print("######################################################################################")
print("# ESTADISTICAS AVANZADAS DE LOS JUGADORES DEL REAL MADRID EN LA TEMPORADA 2022-2023")
print("######################################################################################")
df_estadisticas_avanzadas = procesar_ADVANCED_STATS(df_totales_2022, df_equipo_total, df_totales_rival_2022)
print(df_estadisticas_avanzadas.to_string(index=False, col_space=4))
